changed path of router and added '-' to cocktail-list, profile-list and profile favourites
